functions/remove-background/requirements.txt
pillow==9.5.0
numpy==1.24.3
rembg==2.0.50
